Fixed bugged implementation of google maps api
