+++ src/app/components/driver-contact-modal/driver-contact-modal.component.ts
@@ -25,8 +25,6 @@ export class DriverContactModalComponent implements OnInit {
    */
   ngOnInit() {
 
-    this.getGoogleApi();
-
     this.sleep(2000).then(() => {
      //show drivers on map
      this.showDriversOnMap(sessionStorage.getItem("origin"), sessionStorage.getItem("destination"));
@@ -42,25 +40,7 @@ export class DriverContactModalComponent implements OnInit {
   }
   
 
-/**
- * Appends the google maps api script to the document head.
- */
-getGoogleApi()  {
-    this.http.get(`${environment.loginUri}getGoogleApi`)
-       .subscribe(
-                 (response) => {
-                     //console.log(response);
-                     if(response["googleMapAPIKey"] != undefined){
-                         new Promise((resolve) => {
-                           let script: HTMLScriptElement = document.createElement('script');
-                           script.addEventListener('load', r => resolve());
-                           script.src = `http://maps.googleapis.com/maps/api/js?key=${response["googleMapAPIKey"][0]}`;
-                           //document.head.appendChild(script);      
-                     }); 
-               }    
-           }
-       );
-   }
+
   /**
    * Inititates services from the google maps api needed for displaying the route, and then
    * calls displayRoute()/
